+++ src/IODD.Structure/Structure/ProcessData/ConditionT.cs
@@ -1,3 +1,3 @@
 namespace IOLinkNET.IODD.Structure.ProcessData;
 
-public record ConditionT(string VariableId, byte Subindex, byte Value);
+public record ConditionT(string VariableId, byte Subindex, int Value);
